fix: header tabs and move launch entity to hero
